fixed nits for vs2008
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1938 d60aafaf-7936-0410-ba4d-d307febf7868

+++ src/Midi.cpp
@@ -1951,7 +1951,7 @@ void MidiInPort::exportMidiMessageSequence(MidiMessageSequence* seq,
 	      ev->message.isMidiChannelMetaEvent() &&
 	      (ev->message.getMetaEventLength()>0))
 	    {
-	      line << ev->message.getMetaEventData()[0];
+	      line << (int)(ev->message.getMetaEventData()[0]);
 	    }
 	  else if (ev->message.isTextMetaEvent() ||
 		   ev->message.isTrackNameEvent() )
